+++ lib/primitive/lib/Cardano/Wallet/Primitive/Ledger/Read/Tx.hs
@@ -40,6 +40,9 @@ import Cardano.Read.Ledger.Tx.Outputs
 import Cardano.Read.Ledger.Tx.ScriptValidity
     ( getEraScriptValidity
     )
+import Cardano.Read.Ledger.Tx.Withdrawals
+    ( getEraWithdrawals
+    )
 import Cardano.Wallet.Primitive.Ledger.Read.Tx.Features.CollateralInputs
     ( getCollateralInputs
     )
@@ -70,9 +73,6 @@ import Cardano.Wallet.Read
     )
 import Cardano.Wallet.Read.Eras
     ( eraValue
-    )
-import Cardano.Wallet.Read.Tx.Withdrawals
-    ( getEraWithdrawals
     )
 import Data.Foldable
     ( fold
